# openapi.yaml
openapi: 3.0.3
info:
  title: Mock ERP API (ERPNext Aligned)
  version: v15.0.0
  description: >
    Mock ERP server simulating ERPNext (Frappe Framework) REST API conventions.
    All resource endpoints are available under `/api/resource/{DocType}`.
servers:
  - url: http://localhost:8081/api
    description: Mock ERP Server
components:
  securitySchemes:
    TokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "Format: token {api_key}:{api_secret}"
  schemas:
    Error:
      type: object
      properties:
        exc_type:
          type: string
          example: "DoesNotExistError"
        exception:
          type: string
          example: "frappe.exceptions.DoesNotExistError: Resource not found"
        _server_messages:
          type: string
          example: "[]"

    # Finance Schemas
...
    IntegrationEchoRequest:
      type: object
      additionalProperties: true
      example:
        nested: { value: 7 }
        items: [a, b]
    IntegrationEchoResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/IntegrationEchoRequest'
    IntegrationEchoLastResponse:
      type: object
      required: [data]
      properties:
        data:
          allOf:
            - $ref: '#/components/schemas/IntegrationEchoRequest'
          nullable: true
paths:
...
  /integration/echo:
    post:
      summary: Echo and record one JSON object for integration-test inspection.
      security: [{ TokenAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntegrationEchoRequest'
      responses:
        '200':
          description: The exact received object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationEchoResponse'
        '401':
          description: ERPNext-style authentication error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: The request body is not a JSON object.
  /integration/echo/last:
    get:
      summary: Return the last object received by the echo fixture.
      security: [{ TokenAuth: [] }]
      responses:
        '200':
          description: The last object, or null before the first successful echo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationEchoLastResponse'
        '401':
          description: ERPNext-style authentication error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'